requirements.txt
torch
torchvision
torchaudio
diffusers[torch]
transformers
accelerate
safetensors
streamlit
Pillow
numpy
